Particles endlessly spawn in
Particles are initially added to a pool of available particles. When
particles hit an intake fan, they are added back to the pool, every
spawnRate ms, if a particle is in the pool, it as spawned back in. This
creates a seemingly endless stream of particles, whilst only actually
having a small number to create and simulate