Prototype manual collision detection on fans
By manually implementing collision detection on fans we can go back to a
valid state after going to an invalid state (e.g. Move a fan through
another fan will turn invalid only when we are actually colliding with
the fan), which is something we could not have done using the collision
event listener.

It will also allow us to implement turning invalid state when we move a
fan so that it "hangs" off the side of the case.